+++ packages/boxplot/src/BoxPlot.tsx
@@ -219,6 +219,8 @@ const InnerBoxPlot = <RawDatum extends BoxPlotDatum>({
             medianWidth,
             whiskerWidth,
             whiskerEndWidth,
+            padding,
+            innerPadding,
             isInteractive,
             onClick,
             onMouseEnter,
@@ -236,12 +238,14 @@ const InnerBoxPlot = <RawDatum extends BoxPlotDatum>({
             medianWidth,
             whiskerWidth,
             whiskerEndWidth,
-            getTooltipLabel,
+            padding,
+            innerPadding,
             isInteractive,
             onClick,
             onMouseEnter,
             onMouseLeave,
             tooltip,
+            getTooltipLabel,
             isFocusable,
             boxPlotAriaLabel,
             boxPlotAriaLabelledBy,
